Work in progress. Basic framework for dealing with multiple repetitions. Use a list of tables, one for each dice set, to store rolls and related information. Can be broken down into three parts: 1. Initial rolls. e.g., 3d6. 2. Modifications. e.g. exploding or keep highest. 3. Successes.
Use an S3 class for the list of tables structure, storing other relevant information, like the command, in attributes. R6 might be better for this (particularly for the re-rolling of tables), but then we have to deal with the fact that R6 modifies in place, which is a bit counterintuitive for general R users.

Successes could be pulled out as a method, but that would break the use of command as the primary way to set up the dice rolling parameters.